--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -438,7 +438,6 @@ jobs:
     resource_class: xlarge
     environment:
       EMCC_CORES: 16
-      EMCC_USE_NINJA: 1
     steps:
       - checkout
       - run: