Large system libraries like libc consist of lots of individual objects.

Previously, we'd compile each of them individually, although the processes themselves were parallelised across CPU cores.

However, this resulted in a lot of per-object overhead: create a Python process for Emscripten itself, compile our Python code, execute a bunch of custom logic in said Python code, create another process for Clang with the computed flags, and finally get an object. This overhead is even more pronounced on Windows where (leaving aside details on why) creating a new process is much more expensive than on Linux.

Luckily, Clang, and, by extension, Emscripten, allows to compile lots of source files into object files in the same command invocation / same process, as long as all of them share same compilation flags and target directory (which has to be current working directory).

In this PR I'm leveraging that feature by grouping all the sources within a library by compiler and flags they use. Then, I'm splitting said groups into chunks. Chunk size is currently chosen as `2 * number_of_cores` as a heuristic to strike balance between creating large batches and avoiding workload imbalance where some CPU core might've already finished and stays idle while others are still processing their share of source files. Finally, each chunk is added to the list of commands, and all of those commands are executed in parallel across CPU cores like before. (There is still one exception for commands that need to be singled out, but I won't go into those details here.)

On my laptop with 8 cores, this leads to the following speedups on large system libraries when building `./em++ temp.cpp -s STRICT` where `temp.cpp` is just `int main() {}` and the cache is empty:

| Library Name      | Before (Windows) | After (Windows) | Before (Linux) | After (Linux) | Speedup (Windows) | Speedup (Linux) |
|-------------------|------------------|-----------------|----------------|---------------|-------------------|-----------------|
| libc-debug.a      | 141.7s           | 29.6s           | 53.2s          | 7.7s          | 4.8x              | 7.0x            |
| libcompiler_rt.a  | 22.7s            | 6.6s            | 8.7s           | 1.8s          | 3.5x              | 4.8x            |
| libc++-noexcept.a | 39.5s            | 37.0s           | 20.8s          | 18.6s         | 1.1x              | 1.1x            |