@RReverser 2023-10-30
I'm not seeing those Windows CI failures on Windows locally but I suspect I know why it's happening (seen the same in recent Ninja PR), should be easy to fix.