@RReverser 2023-10-31
> (I didn't even know you can pass in multiple filenames with `-c`...)

It's actually something I only learnt a few days ago myself, and immediately was like "wait a second... I know where this would be useful" 😅