@sbc100 2023-10-31
Nice!   

I guess its a good job that clang doesn't using parallelism internally compile the different source files that its passed, otherwise we would end up with N * N.