@sbc100 2023-10-31
Note that when you have multiple inputs in a single compile command you loose the ability to specify output filenames.   So things like `clang -c dir1/foo.c and dir2/foo.c` won't work since both sources compile to `foo.o`.

Luckily I guess in this case we don't care much about the names of the object files, and we address name collisions as they arise.