@RReverser 2023-10-31
> I guess its a good job that clang doesn't using parallelism internally compile the different source files that its passed, otherwise we would end up with N * N.

I don't think they would do that implicitly, but I actually searched for it and saw a PR where someone was adding parallelism for clang-cl and said they could do it for clang itself via `-j` if there was interest, but that didn't go anywhere.

It would be nice if Clang has such opt-in parallelism since then we wouldn't even need to spawn bunch of processes and could instead pass all the source files in a batch to a single Clang command and let it build those in parallel threads (which are much more lightweight), making builds even faster... Oh well.