@RReverser 2023-10-31
> Note that when you have multiple inputs in a single compile command you loose the ability to specify output filenames. So things like `clang -c dir1/foo.c and dir2/foo.c` won't work since both sources compile to `foo.o`.

Yup, I know, that's why I still have that exemption in code where commands that would result in same object file are split out and handled separately.