@RReverser 2023-10-31
Had to fix one more issue this uncovered (see comment in code), and I think this is good to go.