@sbc100 2023-10-31
I'm surprised the speedup is bigger on linux, given that windows is that one that we expect to see have a higher cost of process creation.