@RReverser 2023-10-31
> I'm surprised the speedup is bigger on linux, given that windows is that one that we expect to see have a higher cost of process creation.

I think this is one of the cases where it should be looked at not so much a multiplier kind of increase (new time / previous time like I did), as absolute reduction of `overhead * number_of_files.`.

So e.g. `libc-debug` which has 1024 files reduced overhead by (141.7-29.6)/1024=0.1 seconds per file on Windows and (53.2-7.7)/1024=0.04 seconds on Linux.

Looking at it this way scales more correctly when we look at other libraries - e.g. for `libcompiler_rt` which has 175 files we get (22.7-6.6)/175=0.09 seconds reduction on Windows and (8.7-1.8)/175=0.04 seconds on Linux, basically same time per file.