@dschuff 2023-10-31
One thing to watch out for on testing here: the CircleCI builder actually uses ninja to prebuild the libraries, so IIUC that's not covering this change. Can you maybe try a commit that removes EMCC_USE_NINJA from the .config.yml and tests that way? It may still be slower than ninja so we may want to revert it, but at least we can test it before landing. The emscripten-releases CI uses this builtin way, so we will get coverage, but not until after it lands here.