@sbc100 2023-10-31
If this method ends up being faster than ninja I wonder if we should just remove the EMCC_USE_NINJA support?