@RReverser 2023-10-31
> One thing to watch out for on testing here: the CircleCI builder actually uses ninja to prebuild the libraries, so IIUC that's not covering this change. Can you maybe try a commit that removes EMCC_USE_NINJA from the .config.yml and tests that way? It may still be slower than ninja so we may want to revert it, but at least we can test it before landing. The emscripten-releases CI uses this builtin way, so we will get coverage, but not until after it lands here.

It's still much faster than Ninja, yes.

I actually initially implemented this change for Ninja as I'm using it locally as well. But I got there only partially because Ninja doesn't have built-in support for batching yet (https://github.com/ninja-build/ninja/issues/1004) and the workaround to find out which files have changed so that only those are batched is rather involved - you need to have task that just collects changed filenames into a temporary file, then chunk that temporary files into batches... it gets complicated.

So I decided to stop there and implement for non-Ninja version of the building script instead, which is both much easier implementation-wise and benefits more users as most didn't opt-in to Ninja yet.

I'll try to remove EMCC_USE_NINJA from config.