@dschuff 2023-10-31
> > One thing to watch out for on testing here: the CircleCI builder actually uses ninja to prebuild the libraries, so IIUC that's not covering this change.
> 
> If this is the case (which it seems to be), I'm a bit surprised some tests failed before due to a change in this non-Ninja path (due to me spawning subprocesses in a different working directory when building system library)...

It only uses ninja for the 'build libraries' step, before tests run But that doesn't build every configuration of every library. IIRC there are still some libraries that get built on-demand during testing runs, and we don't use ninja for that.