@RReverser 2023-10-31
@dschuff Looks like it passes w/o EMCC_USE_NINJA too. CI seems to have a lot of build time variation on commit-to-commit basis though, so it's hard to judge if it's any faster there, but at least it still works.