@RReverser 2023-10-31
One more thing I could probably do is share this logic for ports as well, as I see some ports can also be pretty slow (e.g. libsqlite3 takes 47.57s and libicu takes 38.98s).

But I'd rather leave this for a separate follow-up PR.