@dschuff 2023-10-31
OK great. I think it probably makes sense to check this in without changing the test configuration for now, and we can experiment with that in a separate PR. Assuming Sam is OK with the actual change, I don't even mind pushing it in before the tests finish after you revert the config bit :)