@dschuff 2023-10-31
And yeah I agree that speeding up ports would be great in a followup PR, and that's something I didn't get around to with ninja.