@RReverser 2023-10-31
Ok I'm going to address Sam's comment above first and revert Ninja change. I think I'll still wait for Sam's review / approval before enabling auto-merge though.