@sbc100 2023-10-31
> > And yeah I agree that speeding up ports would be great in a followup PR, and that's something I didn't get around to with ninja.
> 
> Out of curiosity, why do ports even have a different codepath and implementation of building instead of reusing the `Library` class? Are they conceptually different from system libraries?

Mostly just the history of how they evolved I think.

One big difference is that ports are one-per-file (the idea was to allow folks to add new ones by just creating new files in tools/ports.   Whereas system libraries are all declared in the same time.