@RReverser 2023-10-31
> One big difference is that ports are one-per-file (the idea was to allow folks to add new ones by just creating new files in tools/ports.

Okay, but the build process for objects + archive is roughly the same?